Fix document polyfill executed too late
The polyfill MUST be run before react initializes to be effective.
