+++ packages/react/src/render.tsx
@@ -1,3 +1,4 @@
+import './error-polyfill';
 import type {ReactElement} from 'react';
 import type {RemoteRoot} from '@remote-ui/core';
 import type {RootTag} from 'react-reconciler';
@@ -14,24 +15,6 @@ const cache = new WeakMap<
     renderContext: RenderContextDescriptor;
   }
 >();
-
-// Make the react build believe that it is running in a page context,
-// otherwise react rendering errors won't be logged. See:
-// https://github.com/facebook/react/blob/4e6eec69be632c0c0177c5b1c8a70397d92ee181/packages/shared/invokeGuardedCallbackImpl.js#L53-L237
-if (!('window' in globalThis) && !('document' in globalThis)) {
-  class FakeDocument {
-    createEvent(type: string) {
-      return new Event(type);
-    }
-
-    createElement() {
-      return new EventTarget();
-    }
-  }
-
-  globalThis.window = globalThis as any;
-  globalThis.document = new FakeDocument() as any;
-}
 
 // @see https://github.com/facebook/react/blob/993ca533b42756811731f6b7791ae06a35ee6b4d/packages/react-reconciler/src/ReactRootTags.js
 // I think we are a legacy root?
